Fix for #5
Renamed tag-label to "tag-label-blue" and included a "tag-label-red" as
well.  Changed tree icon to the "red icon", when discount not is active.
I included the new image icon in the embeded resource.

I have only a check on the active property. I'm not sure if the active
property change when the discount code expire, otherwise there could
also be a check on the startdate and enddate, so the blue tag icon is
displayed when the discount is active in the webshop and the red icon,
when it isn't active.